backend/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.0
Flask-Migrate==3.1.0
requests==2.26.0
